Deploy IS21 to UAT (master)

+++ backend/api/Areas/Leases/Controllers/SecurityDepositReturnController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using Pims.Api.Areas.Lease.Models.Lease;
 using Pims.Api.Helpers.Exceptions;
 using Pims.Api.Models;
+using Pims.Api.Models.Concepts;
 using Pims.Api.Policies;
 using Pims.Dal.Entities;
 using Pims.Dal.Security;
@@ -48,19 +49,24 @@ namespace Pims.Api.Areas.Lease.Controllers
         /// Adds the specified return deposit to the lease.
         /// </summary>
         /// <returns></returns>
-        [HttpPost("{leaseId:long}/deposit-returns")]
+        [HttpPost("{leaseId:long}/deposits/{depositId:long}/returns")]
         [HasPermission(Permissions.LeaseEdit)]
         [Produces("application/json")]
         [ProducesResponseType(typeof(IEnumerable<LeaseModel>), 200)]
         [SwaggerOperation(Tags = new[] { "lease" })]
-        public IActionResult AddDepositReturn(long leaseId, [FromBody] ParentConcurrencyGuardModel<SecurityDepositReturnModel> addRequest)
+        public IActionResult AddDepositReturn(long leaseId, long depositId, [FromBody] ParentConcurrencyGuardModel<SecurityDepositReturnModel> addRequest)
         {
             if (leaseId != addRequest.ParentId)
             {
                 throw new BadRequestException($"Concurrency parent id mismatch.");
             }
+
+            if (depositId != addRequest.Payload.ParentDepositId)
+            {
+                throw new BadRequestException($"Bad payload id.");
+            }
+
             var depositEntity = _mapper.Map<PimsSecurityDepositReturn>(addRequest.Payload);
-            depositEntity.LeaseId = leaseId;
 
             var updatedLease = _pimsService.SecurityDepositReturnService.AddLeaseDepositReturn(addRequest.ParentId, addRequest.ParentRowVersion, depositEntity);
 
@@ -71,25 +77,24 @@ namespace Pims.Api.Areas.Lease.Controllers
         /// Update the specified return deposit on the passed lease.
         /// </summary>
         /// <returns></returns>
-        [HttpPut("{leaseId:long}/deposit-returns/{depositId:long}")]
+        [HttpPut("{leaseId:long}/deposits/{depositId:long}/returns/{depositReturnId:long}")]
         [HasPermission(Permissions.LeaseEdit)]
         [Produces("application/json")]
         [ProducesResponseType(typeof(IEnumerable<LeaseModel>), 200)]
         [SwaggerOperation(Tags = new[] { "lease" })]
-        public IActionResult UpdateDepositReturn(long leaseId, long depositId, ParentConcurrencyGuardModel<SecurityDepositReturnModel> updateRequest)
+        public IActionResult UpdateDepositReturn(long leaseId, long depositId, long depositReturnId, ParentConcurrencyGuardModel<SecurityDepositReturnModel> updateRequest)
         {
             if (leaseId != updateRequest.ParentId)
             {
                 throw new BadRequestException($"Concurrency parent id mismatch.");
             }
 
-            if (depositId != updateRequest.Payload.Id)
+            if (depositId != updateRequest.Payload.ParentDepositId || depositReturnId != updateRequest.Payload.Id)
             {
                 throw new BadRequestException($"Bad payload id.");
             }
 
             var depositEntity = _mapper.Map<PimsSecurityDepositReturn>(updateRequest.Payload);
-            depositEntity.LeaseId = leaseId;
             var updatedLease = _pimsService.SecurityDepositReturnService.UpdateLeaseDepositReturn(updateRequest.ParentId, updateRequest.ParentRowVersion, depositEntity);
 
             return new JsonResult(_mapper.Map<LeaseModel>(updatedLease));
@@ -99,17 +104,23 @@ namespace Pims.Api.Areas.Lease.Controllers
         /// Delete the specified return deposit on the passed lease.
         /// </summary>
         /// <returns></returns>
-        [HttpDelete("{leaseId:long}/deposit-returns")]
+        [HttpDelete("{leaseId:long}/deposits/{depositId:long}/returns")]
         [HasPermission(Permissions.LeaseEdit)]
         [Produces("application/json")]
         [ProducesResponseType(typeof(IEnumerable<LeaseModel>), 200)]
         [SwaggerOperation(Tags = new[] { "lease" })]
-        public IActionResult DeleteDepositReturn(long leaseId, [FromBody] ParentConcurrencyGuardModel<SecurityDepositReturnModel> deleteRequest)
+        public IActionResult DeleteDepositReturn(long leaseId, long depositId, [FromBody] ParentConcurrencyGuardModel<SecurityDepositReturnModel> deleteRequest)
         {
             if (leaseId != deleteRequest.ParentId)
             {
                 throw new BadRequestException($"Concurrency parent id mismatch.");
             }
+
+            if (depositId != deleteRequest.Payload.ParentDepositId)
+            {
+                throw new BadRequestException($"Bad payload id.");
+            }
+
             var depositEntity = _mapper.Map<PimsSecurityDepositReturn>(deleteRequest.Payload);
             var updatedLease = _pimsService.SecurityDepositReturnService.DeleteLeaseDepositReturn(deleteRequest.ParentId, deleteRequest.ParentRowVersion, depositEntity);
